// src/lib/services/models/InitiateRegistrationResponseChitStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HttpStatusCode } from './HttpStatusCode';
import type { InitiateRegistrationResponse } from './InitiateRegistrationResponse';
export type InitiateRegistrationResponseChitStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: InitiateRegistrationResponse;
    statusCode?: HttpStatusCode;
    errors?: any;
};